Fix ambiguities due to JuliaLang/julia#25261

+++ src/circ_deque.jl
@@ -87,8 +87,8 @@ end
 
 # Iteration via getindex
 @inline Base.start(d::CircularDeque) = 1
-@inline Base.next(d::CircularDeque, i) = (_unsafe_getindex(d, i), i+1)
-@inline Base.done(d::CircularDeque, i) = i == d.n + 1
+@inline Base.next(d::CircularDeque, i::Int) = (_unsafe_getindex(d, i), i+1)
+@inline Base.done(d::CircularDeque, i::Int) = i == d.n + 1
 
 function Base.show(io::IO, D::CircularDeque{T}) where T
     print(io, "CircularDeque{$T}([")
