+++ src/ordered_set.jl
@@ -44,9 +44,9 @@ copy(s::OrderedSet) = union!(similar(s), s)
 empty!(s::OrderedSet{T}) where {T} = (empty!(s.dict); s)
 
 start(s::OrderedSet)       = start(s.dict)
-done(s::OrderedSet, state) = done(s.dict, state)
+done(s::OrderedSet, state::Int) = done(s.dict, state)
 # NOTE: manually optimized to take advantage of OrderedDict representation
-next(s::OrderedSet, i)     = (s.dict.keys[i], i+1)
+next(s::OrderedSet, i::Int)     = (s.dict.keys[i], i+1)
 
 pop!(s::OrderedSet) = pop!(s.dict)[1]
